refactor(Label): remove the interface to index file
