--- src/components/inputs/Label/interfaces/Label.interface.ts
@@ -1,11 +1,0 @@
-import { TypographyLabel } from "../types/Label.type";
-
-export interface ILabelProps {
-  isDisabled?: boolean;
-  isFocused?: boolean;
-  htmlFor: string;
-  isInvalid?: boolean;
-  typo?: TypographyLabel;
-  children?: React.ReactNode;
-  onClick?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-}
